Refactors footer component and style files
