+++ src/app/top/footer/Footer.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './style/footer'
 
 export const Footer = () => {
   return (
