+++ src/app/top/footer/style/footer.scss
@@ -1,0 +1,11 @@
+.footer-wrapper {
+  display: flex;
+  justify-content: center;
+  text-align: center;
+}
+
+@media only screen and (max-width: 420px) {
+  .footer-wrapper {
+    padding: 0 10px 0 10px;
+  }
+}
