+++ src/app/top/style/global.scss
@@ -1,8 +1,3 @@
-/* Global parameters and imports */
-/* html {
-  scroll-behavior: smooth;
-} */
-
 body {
   margin: 0 auto;
   font-family: Arial;
@@ -69,24 +64,6 @@ p {
   flex-direction: column;
   min-height: 100vh;
   width: 100vw;
-}
-
-/* Footer */
-.footer-wrapper {
-  display: flex;
-  justify-content: center;
-  text-align: center;
-}
-
-@media only screen and (max-width: 420px) {
-  .footer-wrapper {
-    padding: 0 10px 0 10px;
-  }
-}
-
-/* Resume */
-.resume-link {
-  margin: 0 0 0 15px;
 }
 
 /* Media query configuration */
